+++ wemake_python_styleguide/visitors/ast/complexity/nested.py
@@ -7,7 +7,7 @@ from wemake_python_styleguide.constants import (
     NESTED_CLASSES_WHITELIST,
     NESTED_FUNCTIONS_WHITELIST,
 )
-from wemake_python_styleguide.types import AnyFunctionDef, AnyNodes
+from wemake_python_styleguide.types import AnyFunctionDef, AnyNodes, final
 from wemake_python_styleguide.violations.best_practices import (
     NestedClassViolation,
     NestedFunctionViolation,
@@ -16,6 +16,7 @@ from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 from wemake_python_styleguide.visitors.decorators import alias
 
 
+@final
 @alias('visit_any_function', (
     'visit_FunctionDef',
     'visit_AsyncFunctionDef',